feat: removing tailwind and adding initial foundations
