+++ packages/react/package.json
@@ -25,7 +25,8 @@
   ],
   "scripts": {
     "start": "storybook dev -p 6006 --no-open",
-    "build": "rimraf ./dist && tsc && cp src/jigsaw.css ./dist/tailwind.css",
+    "build": "rimraf ./dist && npm run build:js",
+    "build:js": "tsc",
     "build:storybook": "storybook build -o dist-storybook",
     "lint": "eslint ."
   },
@@ -40,19 +41,18 @@
     "clsx": "^2.1.1",
     "lucide-react": "^0.494.0",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "tailwind-merge": "^3.2.0"
+    "react-dom": "^19.0.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.22.0",
     "@storybook/addon-essentials": "^8.6.12",
     "@storybook/blocks": "^8.6.12",
     "@storybook/experimental-addon-test": "^8.6.12",
+    "@storybook/manager-api": "^8.6.12",
     "@storybook/react": "^8.6.12",
     "@storybook/react-vite": "^8.6.12",
     "@storybook/test": "^8.6.12",
-    "@tailwindcss/typography": "^0.5.16",
-    "@tailwindcss/vite": "^4.1.4",
+    "@storybook/theming": "^8.6.12",
     "@types/node": "^22.14.1",
     "@types/react": "^19.0.10",
     "@types/react-dom": "^19.0.4",
@@ -66,11 +66,11 @@
     "eslint-plugin-storybook": "^0.12.0",
     "globals": "^16.0.0",
     "lightningcss": "^1.29.3",
+    "lightningcss-cli": "^1.29.3",
     "playwright": "^1.51.1",
     "remark-gfm": "^4.0.1",
     "rimraf": "^6.0.1",
     "storybook": "^8.6.12",
-    "tailwindcss": "^4.1.4",
     "typescript": "~5.7.2",
     "typescript-eslint": "^8.26.1",
     "vite": "^6.3.0",
