+++ packages/react/src/00-foundations/borders/borders.css
@@ -1,0 +1,12 @@
+:root {
+    --border-width-sm: 1px;
+    --border-width-md: 2px;
+    --border-width-lg: 4px;
+    --border-width-xl: 8px;
+
+    --radius-sm: 4px;
+    --radius-md: 6px;
+    --radius-lg: 8px;
+    --radius-xl: 10px;
+
+}
