+++ packages/react/src/00-foundations/breakpoints/breakpoints.css
@@ -1,0 +1,6 @@
+:root {
+    --breakpoint-sm: 576px;
+    --breakpoint-md: 768px;
+    --breakpoint-lg: 992px;
+    --breakpoint-xl: 1200px;
+}
