+++ packages/react/src/00-foundations/colours/colours.mdx
@@ -2,9 +2,7 @@ import { Meta, ColorPalette, ColorItem } from "@storybook/blocks";
 
 <Meta title="Foundations/Colours" />
 
-# Colours
-
-## All Colours
+# All Colours
 
 <ColorPalette>
   <ColorItem
@@ -181,103 +179,6 @@ import { Meta, ColorPalette, ColorItem } from "@storybook/blocks";
       "yellow-700": "#b88c1b",
       "yellow-800": "",
       "yellow-900": "",
-    }}
-  />
-</ColorPalette>
-
-## Theme
-
-<ColorPalette>
-  <ColorItem
-    title="Background"
-    subtitle="Primary brand"
-    colors={{
-      "background 3 (navy-900)": "#121419",
-      "background 2(navy-800)": "#191B23",
-      "background (navy-500)": "#232530",
-      "foreground 1 (navy-300)": "#2D303E",
-      "foreground 2 (navy-100)": "#383B4C",
-    }}
-  />
-  <ColorItem
-    title="Teal"
-    subtitle="Primary brand"
-    colors={{
-      "text (mono-100)": "#eee",
-      "background (teal-500)": "#0a766d",
-      "background hover (teal-700)": "#07564f",
-    }}
-  />
-  <ColorItem
-    title="Grey"
-    subtitle="Primary brand"
-    colors={{
-      "text (mono-100)": "#eee",
-      "background (grey-700)": "#4D5868",
-      "background hover (grey-900)": "#374151",
-    }}
-  />
-  <ColorItem
-    title="Red"
-    subtitle="Primary brand"
-    colors={{
-      "text (mono-100)": "#eee",
-      "background (red-500)": "#c92e2e",
-      "background hover (red-700)": "#a22525",
-    }}
-  />
-  <ColorItem
-    title="Green"
-    subtitle="Primary brand"
-    colors={{
-      "text (mono-100)": "#eee",
-      "background (green-500)": "#377845",
-      "background hover (green-700)": "#2b5d36",
-    }}
-  />
-  <ColorItem
-    title="Blue"
-    subtitle="Primary brand"
-    colors={{
-      "text (mono-100)": "#eee",
-      "background (blue-500)": "#18719e",
-      "background hover (blue-700)": "#13597c",
-    }}
-  />
-  <ColorItem
-    title="Purple"
-    subtitle="Primary brand"
-    colors={{
-      "text (mono-100)": "#eee",
-      "background (purple-500)": "#7855c8",
-      "background hover (purple-700)": "#6245a4",
-    }}
-  />
-  <ColorItem
-    title="Pink"
-    subtitle="Primary brand"
-    colors={{
-      "text (mono-100)": "#eee",
-      "background (pink-500)": "#b92971",
-      "background hover (pink-700)": "#95205a",
-    }}
-  />
-  <ColorItem
-    title="Orange"
-    subtitle="Primary brand"
-    colors={{
-      "text (mono-900)": "#111",
-      "background (orange-400)": "#e1762b",
-      "background hover (orange-700)": "#a4551f",
-    }}
-  />
-  <ColorItem
-    title="Yellow"
-    subtitle="Primary brand"
-    colors={{
-      "text (mono-900)": "#111",
-      "background (yellow-400)": "#fbbf24",
-      "background hover (yellow-700)": "#b88c1b",
     }}
   />
 </ColorPalette>
