+++ packages/react/src/00-foundations/colours/theme.css
@@ -1,0 +1,3 @@
+:root {
+    --theme-background: var(--color-navy-500)
+}
