+++ packages/react/src/01-atoms/prose/prose.css
@@ -1,0 +1,3 @@
+.prose {
+
+}
