+++ packages/react/src/jigsaw.ts
@@ -1,5 +1,3 @@
 
-export * from "./utils/twclsx.js"
-
 export * from "./01-atoms/button/button.js"
 export * from "./01-atoms/prose/prose.js"
