+++ packages/react/vite.config.ts
@@ -1,7 +1,6 @@
 import { defineConfig } from 'vite'
 
 import react from '@vitejs/plugin-react'
-import tailwindcss from '@tailwindcss/vite'
 
 // https://vite.dev/config/
 export default defineConfig({
@@ -17,7 +16,6 @@ export default defineConfig({
     cssMinify: "lightningcss"
   },
   plugins: [
-    react(),
-    tailwindcss(),
+    react()
   ]
 })
